requirements.txt
Pillow==7.2.0
pytesseract==0.3.6
win32gui==0.6
telegram-send==0.25
pynput==1.6.8
json==2.0.9
